basic_rest_api_project/requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.4.0
djangorestframework==3.15.1
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2023.3
